refactored folder structure to reflect inherited widget change
